Moved to_string converters to convert_to_string namespace.
