+++ src/instrumentation/converter/convert_to_string/double_to_string.cc
@@ -1,4 +1,4 @@
-#include "instrumentation/converter/to_string/double_to_string.h"
+#include "instrumentation/converter/convert_to_string/double_to_string.h"
 
 #include "instrumentation/converter/factory.hpp"
 
@@ -11,7 +11,7 @@ namespace instrumentation {
 
 namespace converter {
 
-namespace to_string {
+namespace convert_to_string {
 
 namespace  {
 using OutputTerm = DoubleToStringOutputTerm;
@@ -46,7 +46,7 @@ bool DoubleToString::is_registered_ =
           return return_ptr;
         });
 
-} // namespace to_string
+} // namespace convert_to_string
 
 } // namespace converter
