+++ src/instrumentation/converter/convert_to_string/tests/convergence_to_string_test.cc
@@ -1,4 +1,4 @@
-#include "instrumentation/converter/to_string/convergence_to_string.h"
+#include "instrumentation/converter/convert_to_string/convergence_to_string.h"
 
 #include <iomanip>
 #include <sstream>
@@ -13,7 +13,7 @@ using namespace bart;
 
 class InstrumentationConverterConvergenceToStringTest : public ::testing::Test {
  public:
-  using ConverterType = instrumentation::converter::to_string::ConvergenceToString;
+  using ConverterType = instrumentation::converter::convert_to_string::ConvergenceToString;
   convergence::Status test_status;
   std::string GetExpectedOutput(convergence::Status&, ConverterType&,
                                 std::string) const;
@@ -88,7 +88,7 @@ std::string InstrumentationConverterConvergenceToStringTest::GetExpectedOutput(
 }
 
 TEST_F(InstrumentationConverterConvergenceToStringTest, Constructor) {
-  using ConverterType = instrumentation::converter::to_string::ConvergenceToString;
+  using ConverterType = instrumentation::converter::convert_to_string::ConvergenceToString;
   EXPECT_NO_THROW({
     ConverterType test_converter;
   });
