+++ src/instrumentation/converter/convert_to_string/to_string_converter.h
@@ -15,7 +15,7 @@ namespace instrumentation {
 
 namespace converter {
 
-namespace to_string {
+namespace convert_to_string {
 
 template <typename InputType, typename OutputTerm>
 class ToStringConverter : public ConverterI<InputType, std::string> {
@@ -51,7 +51,7 @@ class ToStringConverter : public ConverterI<InputType, std::string> {
   OutputTermToStringMap output_term_to_string_map_;
 };
 
-} // namespace to_string
+} // namespace convert_to_string
 
 } // namespace converter
